Add files via upload
Every hospital or clinic should have a system in place for handling electronic health records. This system maintains a record of all pertinent patient information, including name, diagnosis of sickness, contact information (phone number, email address, address, etc.), and diagnosis and prescription information for clinical care.  It also keeps a list of doctors with information about each one, including name, specialization, and phone and email numbers. Additionally, it logs information about the exam room and materials. The system also keeps track of information concerning insurance, such as the provider and plan specifics. Last but not least, visit information upkeep information pertaining to visit date and time.This project is an extension of project 1 on which we use an application programming interface using stored procedures that allows an EMR system to work without having a database and writing sql queries. I have used python programming language to implement. This model also provides features like authentication on authorized people who are supposed to use the database and certain constraints are implemented so that only doctors can update prescriptions. And also features like saving the the history every time a record is accessed.

Functional Requirements specification : This system should be able to create efficient information management system for an student clinic with use of API so no need to use the work bench . Here clinic uses a database to store patient information, visit history, insurance details, care details, prescriptions, exam room details and supplies , billing records and Auth table is add for authentication. Also, the system should make it possible to manage medications, schedule visits, and handle billing procedures effectively. The main objective is to develop a system that uses api to intract with the user without use of workbench for that I have chosen visual studio to work and the system is all efficient to store the audits if any changes are made.
New tables added to the database to satisfy the requirement specified in the porject I have taken the new table related to the audit to store the details of the changes. All the tables are specifiend in functional requirement